move ggemm resolution to megatron/core/inference
