+++ megatron/core/transformer/moe/experts.py
@@ -1,7 +1,6 @@
 # Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 from __future__ import annotations
 
-import enum
 import logging
 from collections.abc import Callable
 from copy import deepcopy
@@ -59,50 +58,12 @@ try:
 except ImportError:
     HAVE_FLASHINFER = False
 
+from megatron.core.inference.moe import (
+    InferenceGroupedGemmBackend,
+    resolve_inference_grouped_gemm_backend,
+)
+
 logger = logging.getLogger(__name__)
-
-
-class InferenceGroupedGemmBackend(enum.Enum):
-    """Resolved backend for grouped GEMM operations during inference."""
-
-    FLASHINFER = "flashinfer"
-    TORCH = "torch"
-    TE = "te"
-
-
-def resolve_inference_grouped_gemm_backend(
-    backend: str,
-    is_cuda_graphed: bool,
-) -> InferenceGroupedGemmBackend:
-    """Resolve the grouped GEMM backend to use for the current iteration.
-
-    Prerequisites are validated at init time in MoELayer; this function
-    simply maps (backend, is_cuda_graphed) to the concrete backend enum.
-
-    Args:
-        backend: One of 'auto', 'torch', 'te'.
-        is_cuda_graphed: Whether this is a CUDA-graphed iteration.
-
-    Returns:
-        An InferenceGroupedGemmBackend enum value.
-    """
-    if backend == 'auto':
-        if is_cuda_graphed:
-            return InferenceGroupedGemmBackend.FLASHINFER
-        else:
-            if hasattr(torch.nn.functional, 'grouped_mm'):
-                return InferenceGroupedGemmBackend.TORCH
-            else:
-                return InferenceGroupedGemmBackend.TE
-    elif backend == 'torch':
-        return InferenceGroupedGemmBackend.TORCH
-    elif backend == 'te':
-        return InferenceGroupedGemmBackend.TE
-    else:
-        raise ValueError(
-            f"Unknown inference_grouped_gemm_backend: '{backend}'. "
-            "Must be 'auto', 'torch', or 'te'."
-        )
 
 
 class GroupedLinearFc1Interface(Protocol):
